requirements.txt
beautifulsoup4==4.7.0
boto3==1.9.71
Flask==1.1.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.5
keyboard==0.13.3
pgi==0.0.11.2
psutil==5.4.7
PyAudio==0.2.11
pynput==1.4
pyperclip==1.7.0
python-telegram-bot==11.1.0
selenium==3.141.0
six==1.11.0
spotipy==2.4.4
tabulate==0.8.3
wikipedia==1.4.0
PyYAML==5.1
requests==2.21.0
grpcio==1.17.1
